Refactor code for improved clarity and performance
- Updated memory management in CacheShard to use correct value size.
- Simplified value mapping in SstFile reader.
- Changed key recording logic in SparseIndexWriter to use is_multiple_of method.
- Streamlined error handling in MockCloudBackend for better readability.
- Added Default derive for EvictionStrategy and DurabilityPolicy enums.
- Cleaned up WAL module documentation for clarity and consistency.
- Optimized sync method in FsWalWriter by removing unnecessary mutable reference.
- Introduced Default implementation for RecoveryStats struct.
- Addressed various Clippy warnings related to documentation formatting and code style.

+++ src/compaction/strategy.rs
@@ -138,7 +138,7 @@ impl Compactor {
             0 => self.config.l0_compaction_threshold,
             1 => self.config.l1_target_size,
             _ => {
-                let exp = (level - 1) as u32;
+                let exp = level - 1;
                 self.config
                     .l1_target_size
                     .saturating_mul(self.config.level_multiplier.saturating_pow(exp))
